third_party/WebKit/LayoutTests/fast/forms/calendar-picker/calendar-picker-key-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="2000-01-02">
<script>
description("Tests if calendar picker key bindings work as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('date'), test1);

function test1() {
    debug('Check that page popup exists.');
    shouldBeEqualToString('popupWindow.pagePopupController.toString()', '[object PagePopupController]');

    shouldBeFalse('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-02');

    debug('Check that arrow keys work properly even when going between weeks and months.');

    debug('press up');
    eventSender.keyDown('ArrowUp');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug('press up');
    eventSender.keyDown('ArrowUp');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-19');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-19');

    debug('press down');
    eventSender.keyDown('ArrowDown');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug('press down');
    eventSender.keyDown('ArrowDown');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-02');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-02');

    debug('press up');
    eventSender.keyDown('ArrowUp');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    // Move from top left of January 2000 to bottom right of December 1999.
    debug('press left');
    eventSender.keyDown('ArrowLeft');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-25');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-25');

    debug('press right');
    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press PageDown");
    eventSender.keyDown('PageDown');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press PageUp");
    eventSender.keyDown('PageUp');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press m");
    eventSender.keyDown('m');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '2000-01');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '2000-01-02');
    shouldBeEqualToString('highlightedValue()', '2000-01-23');
    shouldBeEqualToString('highlightedDayCells()', '2000-01-23');

    debug("press shift + m");
    eventSender.keyDown('M');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press y");
    eventSender.keyDown('y');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '2000-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2000-12-24');
    shouldBeEqualToString('highlightedDayCells()', '2000-12-24');

    debug("press shift + y");
    eventSender.keyDown('Y');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug("press d");
    eventSender.keyDown('d');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '2009-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '2009-12-27');
    shouldBeEqualToString('highlightedDayCells()', '2009-12-27');

    debug("press shift + d");
    eventSender.keyDown('D');
    shouldBeTrue('isCalendarTableScrollingWithAnimation()');
    shouldBeEqualToString('currentMonth()', '1999-12');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '1999-12-26');

    debug('Check if tabbing works.');
    shouldBeEqualToString('focusedElement()', '.list-view calendar-table-view');
    eventSender.keyDown('Tab', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.calendar-navigation-button');
    eventSender.keyDown('Tab', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.calendar-navigation-button today-button');
    eventSender.keyDown('Tab', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.calendar-navigation-button');
    eventSender.keyDown('Tab', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.month-popup-button');

    debug('open the month popup');
    eventSender.keyDown('Enter');
    shouldBeTrue('popupWindow.global.picker.monthPopupView.isVisible');
    shouldBeEqualToString('highlightedMonthButton()', '1999-12');

    debug('TAB shouldn\'t change focus');
    shouldBeEqualToString('focusedElement()', '.list-view year-list-view');
    eventSender.keyDown('Tab');
    // With the MockPagePopup, an elemenet not in the popup is focused.
    shouldBeTrue('focusedElement().indexOf("month-button") < 0');
    shouldBeTrue('focusedElement().indexOf("list-cell") < 0');
    eventSender.keyDown('Tab', ['shiftKey']);
    shouldBeEqualToString('focusedElement()', '.list-view year-list-view');

    debug("press down");
    eventSender.keyDown('ArrowDown');
    shouldBeEqualToString('highlightedMonthButton()', '2000-04');

    debug("press right");
    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('highlightedMonthButton()', '2000-05');

    debug("press right");
    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('highlightedMonthButton()', '2000-06');

    debug("press right");
    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('highlightedMonthButton()', '2000-07');

    debug("press right");
    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('highlightedMonthButton()', '2000-08');

    debug("press right");
    eventSender.keyDown('ArrowRight');
    shouldBeEqualToString('highlightedMonthButton()', '2000-09');

    debug("press left");
    eventSender.keyDown('ArrowLeft');
    shouldBeEqualToString('highlightedMonthButton()', '2000-08');

    debug("press up");
    eventSender.keyDown('ArrowUp');
    shouldBeEqualToString('highlightedMonthButton()', '2000-04');

    debug("press PageDown");
    eventSender.keyDown('PageDown');
    shouldBeEqualToString('highlightedMonthButton()', '2001-04');

    debug("press PageUp");
    eventSender.keyDown('PageUp');
    shouldBeEqualToString('highlightedMonthButton()', '2000-04');

    debug('select month in the month popup');
    eventSender.keyDown('Enter');
    shouldBeFalse('popupWindow.global.picker.monthPopupView.isVisible');
    shouldBeEqualToString('currentMonth()', '2000-04');
    shouldBeEqualToString('selectedValue()', '2000-01-02');
    shouldBeEqualToString('selectedDayCells()', '');
    shouldBeEqualToString('highlightedValue()', '1999-12-26');
    shouldBeEqualToString('highlightedDayCells()', '');

    debug("focus calendar table");
    popupWindow.global.picker.calendarTableView.element.focus();

    setNoCloseOnCommit();
    // Test twice in case midnight.
    debug("press t");
    shouldBeTrue('testToday() || testToday()');

    debug("press esc");
    eventSender.keyDown('Escape');
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');

    finishJSTest();
}

function testToday() {
    eventSender.keyDown('t');
    shouldBeTrueQuiet('isCalendarTableScrollingWithAnimation()');
    var now = new Date();
    var expectedDate = serializeDate(now.getFullYear(), now.getMonth(), now.getDate());
    var expectedMonth = (new popupWindow.Month(now.getFullYear(), now.getMonth())).toString();
    return selectedValue() === expectedDate &&
           highlightedValue() === expectedDate &&
           currentMonth() === expectedMonth &&
           document.getElementById("date").value === expectedDate;
}

function serializeDate(year, month, day) {
    var yearString = String(year);
    if (yearString.length < 4)
        yearString = ("000" + yearString).substr(-4, 4);
    return yearString + "-" + ("0" + (month + 1)).substr(-2, 2) + "-" + ("0" + day).substr(-2, 2);
}

function focusedElement() {
    var element = popupWindow.document.activeElement;
    if (!element)
        return null;
    var identifier = "." + element.className;
    if (element.value)
        identifier += "[value=" + element.value + "]";
    if (identifier == ".")
        return elementIdentifier(element);
    return identifier;
}

function elementIdentifier(element) {
    var identifier = element.parentElement ? (elementIdentifier(element.parentElement) + " ") : "";
    identifier += element.tagName;
    if (element.className)
        identifier += "." + element.className.replace(/ +/g, ".");
    if (element.value)
        identifier += "[value=" + element.value + "]";
    return identifier;
}

</script>
</body>
</html>
